oficios-automation/requirements.txt
google-cloud-firestore>=2.14.0
google-cloud-pubsub>=2.18.0
google-cloud-vision>=3.4.0
google-cloud-aiplatform>=1.38.0
groq>=0.4.0
pydantic>=2.5.0
python-dateutil>=2.8.2
firebase-admin>=6.3.0
PyJWT>=2.8.0
openai>=1.3.0
numpy>=1.24.0
scipy>=1.11.0

